Refactor sign on method components and folder structure
