+++ apps/admin-portal/src/components/applications/sign-on-methods/script-based-flow/index.ts
@@ -16,4 +16,4 @@
  * under the License.
  */
 
-export * from "./sign-on-methods";
+export * from "./script-based-flow";
