+++ apps/admin-portal/src/components/applications/sign-on-methods/script-based-flow/script-based-flow.tsx
@@ -64,7 +64,7 @@ interface AdaptiveScriptsPropsInterface {
  * @param {AdaptiveScriptsPropsInterface} props - Props injected to the component.
  * @return {ReactElement}
  */
-export const AdaptiveScripts: FunctionComponent<AdaptiveScriptsPropsInterface> = (
+export const ScriptBasedFlow: FunctionComponent<AdaptiveScriptsPropsInterface> = (
     props: AdaptiveScriptsPropsInterface
 ): ReactElement => {
