+++ apps/admin-portal/src/components/applications/sign-on-methods/script-based-flow/script-templates-side-panel.tsx
