+++ apps/admin-portal/src/components/applications/sign-on-methods/step-based-flow/index.ts
@@ -16,4 +16,4 @@
  * under the License.
  */
 
-export * from "./adaptive-scripts";
+export * from "./step-based-flow";
